backend/requirements.txt
# ==================== FRAMEWORK WEB ====================
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# ==================== GOOGLE APIS Y SHEETS ====================
gspread==6.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
httplib2==0.22.0

# ==================== UTILIDADES Y CONFIGURACIÓN ====================
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# ==================== MANEJO DE DATOS Y FECHAS ====================
pytz==2023.3
six==1.16.0

# ==================== SEGURIDAD Y SERIALIZACIÓN ====================
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0

# ==================== LOGGING Y MONITOREO (OPCIONAL PERO RECOMENDADO) ====================
python-json-logger==2.0.7

# ==================== TESTING Y DESARROLLO (OPCIONAL) ====================
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# ==================== UTILIDADES ADICIONALES ====================
cachetools==5.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9